Initial State management
Basic state manager class. All states are stored in a list and are
retrieved when the user wants to switch state (i.e starting a new game).
States can also be loaded and unloaded in real time (for example, the
PlayGameState will need to be unloaded to clear any data from a previous
game).